Handle double values in greater/smaller stream matcher.

+++ graylog2-server/src/main/java/org/graylog2/streams/matchers/SmallerMatcher.java
@@ -19,7 +19,7 @@ package org.graylog2.streams.matchers;
 import org.graylog2.plugin.Message;
 import org.graylog2.plugin.streams.StreamRule;
 
-import static org.graylog2.plugin.Tools.getInt;
+import static org.graylog2.plugin.Tools.getDouble;
 
 /**
  * @author Lennart Koopmann <lennart@socketfeed.com>
@@ -28,13 +28,13 @@ public class SmallerMatcher implements StreamRuleMatcher {
 
 	@Override
 	public boolean match(Message msg, StreamRule rule) {
-        Integer msgVal = getInt(msg.getField(rule.getField()));
+        Double msgVal = getDouble(msg.getField(rule.getField()));
 
         if (msgVal == null) {
             return false;
         }
 
-        Integer ruleVal = getInt(rule.getValue());
+        Double ruleVal = getDouble(rule.getValue());
         if (ruleVal == null) {
             return false;
         }
